requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
qrcode
Pillow
Werkzeug
pandas
numpy
matplotlib
seaborn 